Remove bus stop generation.
